+++ R/developer.R
@@ -144,6 +144,11 @@
 #' `stage` argument. Can be used by functions that interact with recipes to 
 #' verify data is correct before passing it to `prep()` and `bake()`.
 #' 
+#' [recipes_ptype_validate()] checks that a recipe and its data are compatible
+#' using information extracted using [recipes_ptype()]. Can be used by functions 
+#' that interact with recipes to verify data is correct before passing it to 
+#' `prep()` and `bake()`.
+#' 
 #' [detect_step()] returns a logical indicator to determine if a given step or
 #' check is included in a recipe.
 #'