+++ R/ptype.R
@@ -46,7 +46,7 @@
 #' @return A zero row tibble.
 #' @keywords internal
 #'
-#' @seealso [developer_functions]
+#' @seealso [developer_functions] [recipes_ptype_validate]
 #' 
 #' @examples
 #' training <- tibble(
@@ -111,3 +111,107 @@ recipes_ptype <- function(x, ..., stage = "prep") {
   
   ptype
 }
+
+#' Validate prototype of recipe object
+#'
+#' This helper function validates a dataframe against the ptype of a recipe.
+#'
+#' @param x A `recipe` object.
+#' @param new_data A data.frame. To be patched aganist ptype of `x`.
+#' @param ... currently not used.
+#' @param stage A single character. Must be one of `"prep"` or `"bake"`. See
+#'   details for more. Defaults to `"prep"`.
+#' @param call The execution environment of a currently running function, e.g.
+#'   `caller_env()`. The function will be mentioned in error messages as the
+#'   source of the error. See the call argument of [rlang::abort()] for more
+#'   information.
+#' 
+#' @return Nothing or an error.
+#' @keywords internal
+#'
+#' @seealso [developer_functions] [recipes_ptype]
+#' 
+#' @examples
+#' rec <- recipe(mpg ~ disp, data = mtcars)
+#' 
+#' recipes_ptype_validate(rec, mtcars)
+#' 
+#' @export
+recipes_ptype_validate <- function(x,
+                                   new_data,
+                                   ...,
+                                   stage = "prep", 
+                                   call = rlang::caller_env()) {
+  old_ptype <- recipes_ptype(x, stage = stage)
+  col_names <- names(old_ptype)
+
+  new_ptype <- vctrs::vec_ptype(new_data)
+
+  if (!all(col_names %in% names(new_ptype))) {
+    offenders <- col_names[!col_names %in% names(new_ptype)]
+    cli::cli_abort(
+      "Not all variables in the recipe are present in the supplied training \\
+      set: {.var {offenders}}.",
+      call = call
+    )
+  }
+
+  new_ptype <- new_ptype[col_names]
+
+  if (!identical(lapply(old_ptype, class), lapply(new_ptype, class))) {
+    old_classes <- lapply(old_ptype, class)
+    new_classes <- lapply(new_ptype, class)
+
+    offenders <- purrr::map2_lgl(old_classes, new_classes, identical)
+    offenders <- col_names[!offenders]
+
+    msg <- c("x" = "{cli::qty(offenders)} The following variable{?s} \\
+                   {?has/have} the wrong class:")
+
+    # Use `paste0()` rather than typical glue syntax to intentionally duplicate messages
+    col_msg <- paste0(
+      "{.var {offenders[",
+      seq_along(offenders),
+      "]}} must have class {.cls {new_classes[offenders[",
+      seq_along(offenders),
+      "]]}}, not {.cls {old_classes[offenders[",
+      seq_along(offenders),
+      "]]}}."
+    )
+    names(col_msg) <- rep("*", length(col_msg))
+    msg <- c(msg, col_msg)
+
+    cli::cli_abort(msg, call = call)
+  }
+    
+  if (!identical(lapply(old_ptype, attributes), lapply(new_ptype, attributes))) {
+    old_attributes <- lapply(old_ptype, attributes)
+    new_attributes <- lapply(new_ptype, attributes)
+  
+    offenders <- purrr::map2_lgl(old_attributes, new_attributes, identical)
+    offenders <- col_names[!offenders]
+      
+    msg <- c("x" = "{cli::qty(offenders)} The following variable{?s} \\
+    {?has/have} the wrong attributes: {.var {offenders}}.")
+
+    if (any(map_lgl(old_ptype[offenders], is.factor))) {
+      offenders_fct <- map_lgl(old_ptype[offenders], is.factor)
+      offenders_fct <- names(offenders_fct)
+    
+      fct_msg <- c(
+        "*" = "The factor levels of {.var {offenders_fct}} don't match."
+      )
+      msg <- c(msg, fct_msg)
+    }
+
+    msg <- c(
+      msg, 
+      "Run {.code lapply(recipes_ptype(rec), attributes)} to see expected \\
+      attributes. For {.code rec} being the name of your recipe."
+    )
+  
+    cli::cli_abort(msg, call = call)
+  }
+      
+  invisible()
+}